Daniel

+++ BeerHound/src/main/java/com/skilldistillery/beerhound/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.skilldistillery.beerhound.controllers;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,14 +29,14 @@ public class UserController {
 	}
 	
 	@RequestMapping(path="getUser.do", params = "id")
-	public String getUserById(Model model, Integer id) {
+	public String getUserById(Model model, Integer id, HttpSession session) {
 		User user = userDao.getUserById(id);
 		model.addAttribute("user", user);
 		return "user/user";
 	}
 	
 	@RequestMapping(path="getUser.do", params = "email")
-	public String getUserByEmail(Model model, String email) {
+	public String getUserByEmail(Model model, String email, HttpSession session) {
 		User user = userDao.getUserByEmail(email);
 		model.addAttribute("user", user);
 		//WEB-INF/user/user.jsp
@@ -50,7 +51,7 @@ public class UserController {
 	}
 	
 	@RequestMapping(path="createUser.do", method=RequestMethod.POST)
-	public String createUser(Model model, @Valid User user, Errors errors) {
+	public String createUser(Model model, @Valid User user, Errors errors, HttpSession session) {
 		if (errors.hasErrors()) {
 			return "user/register";
 		}
@@ -58,14 +59,28 @@ public class UserController {
 			errors.rejectValue("email", "error.email", user.getEmail() + " is already registered");
 			return "user/register";
 		}
-		model.addAttribute("user", userDao.createUser(user));
+		User registeredUser = userDao.createUser(user);
+		model.addAttribute("user", registeredUser);
+		session.setAttribute("loginUser", registeredUser);
 		return "user/user";
 	}
 	
 	@RequestMapping(path="deleteUser.do")
-	public String delete(Model model, Integer id) {
-		
-		boolean wasDeleted = userDao.deleteUser(id);
+	public String delete(Model model, Integer id, HttpSession session) {
+		User loggedInUser = (User) session.getAttribute("loginUser");
+		User userToDelete = userDao.getUserById(id);
+		boolean wasDeleted = false;
+		// if not logged in, cannot delete
+		if (loggedInUser == null || userToDelete == null) {
+			model.addAttribute("deleted", wasDeleted);
+			return "user/delete";
+		}
+		// if not admin and not same user, also cannot delete
+		if (loggedInUser.getRole() != "ADMIN" && loggedInUser.getId() != userToDelete.getId()) {
+			model.addAttribute("deleted", wasDeleted);
+			return "user/delete";
+		}
+		wasDeleted = userDao.deleteUser(id);
 		model.addAttribute("deleted", wasDeleted);
 		if (wasDeleted) {
 			return "user/delete";
@@ -73,6 +88,47 @@ public class UserController {
 			model.addAttribute("user", userDao.getUserById(id));
 			return "user/delete";
 		}
+	}
+	
+	
+	@RequestMapping(path="login", method=RequestMethod.GET)
+	public String login(Model model) {
+		model.addAttribute("user", new User());
+		return "user/login";
+	}
+	
+	@RequestMapping(path="login.do", method=RequestMethod.POST)
+	public String doLogin(Model model, @Valid User user, Errors errors, HttpSession session) {
+		
+		
+		User loggedInUser = userDao.login(user);
+		if (loggedInUser == null) {
+			errors.reject("Username or Password Incorrect");
+		}		
+		if (errors.getErrorCount() != 0) {
+			return "user/login";
+		}
+		session.setAttribute("loginUser", loggedInUser);
+		model.addAttribute("user", loggedInUser);
+		return "user/user";
+	}
+	
+	@RequestMapping(path="updateUser")
+	public String update(Model model, HttpSession session, Integer id) {
+		model.addAttribute("user", userDao.getUserById(id));
+		return "user/update";
+	}
+	
+	@RequestMapping(path="updateUser.do", method=RequestMethod.POST)
+	public String doUpdate(Model model, @Valid User user, Errors errors, HttpSession session) {
+		if (errors.hasErrors()) {
+			model.addAttribute("user", user);
+			return "user/update";
+		}
+		user = userDao.updateUser(user);
+		session.setAttribute("loginUser", user);
+		model.addAttribute("user", user);
+		return "user/user";
 	}
 
 }
